+++ clients/privacy-center/common/geolocation.ts
@@ -1,11 +1,15 @@
 import type { NextApiRequest } from "next";
 import { UserGeolocation } from "fides-js";
 
-// Regex to validate a location string, which must:
-// 1) Start with a 2-3 character country code (e.g. "US")
-// 2) Optionally end with a 2-3 character region code (e.g. "CA")
+// Regex to validate an ISO-3166 location string, which must:
+// 1) Start with a 2-3 letter country code (e.g. "US", "USA", "EEA")
+// 2) Optionally end with a 1-3 alphanumeric character region code (e.g. "CA", "123", "X")
 // 3) Separated by a dash (e.g. "US-CA")
-const VALID_ISO_3166_LOCATION_REGEX = /^\w{2,3}(-\w{2,3})?$/;
+const VALID_ISO_3166_LOCATION_REGEX = /^[a-z]{2,3}(-[a-z0-9]{1,3})?$/i;
+
+// Regex to validate a standalone ISO-3166-2 region code, which must be a 1-3
+// alphanumeric character region code (e.g. "CA", "123", "X")
+const VALID_ISO_3166_2_REGION_REGEX = /^[a-z0-9]{1,3}?$/i;
 
 // Constants for the supported CloudFront geolocation headers
 // (see https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/adding-cloudfront-headers.html#cloudfront-headers-viewer-location)
@@ -50,7 +54,12 @@ export const lookupGeolocation = async (
     geolocation = country;
     if (typeof req.headers[CLOUDFRONT_HEADER_REGION] === "string") {
       [region] = req.headers[CLOUDFRONT_HEADER_REGION].split(",");
-      geolocation = `${country}-${region}`;
+      // Check if the region header is valid; otherwise discard (it's optional!)
+      if (VALID_ISO_3166_2_REGION_REGEX.test(region)) {
+        geolocation = `${country}-${region}`;
+      } else {
+        region = undefined;
+      }
     }
     if (VALID_ISO_3166_LOCATION_REGEX.test(geolocation)) {
       return {